# openapi.yaml
openapi: 3.1.0
info:
  title: Gatekeeper
  description: >
    API gateway on Cloudflare Workers with AWS IAM-style policy authorization.
    Fronts Cloudflare's cache purge API and S3-compatible R2 storage with
    per-key/credential policies, token-bucket rate limiting, request collapsing,
    and D1-backed analytics. The S3 proxy provides fine-grained object-level
    access control that R2's native tokens cannot — key-prefix restrictions,
    per-operation permissions, cross-bucket mixed policies.
  version: 3.0.0
  contact:
    name: Erfi Anugrah
servers:
  - url: https://gate.erfi.io
    description: Production

# ─── Security schemes ──────────────────────────────────────────────────────
security: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      description: >
        Gateway API key (prefix `gw_`). Issued via `POST /admin/keys`.
        Each key has an attached IAM policy that controls which actions,
        resources, and conditions it can access.

    AdminKeyAuth:
      type: apiKey
      in: header
      name: X-Admin-Key
      description: >
        Shared admin secret for CLI and automation. Falls back when
        Cloudflare Access is not configured.

    CloudflareAccess:
      type: apiKey
      in: header
      name: Cf-Access-Jwt-Assertion
      description: >
        Cloudflare Access JWT. Injected automatically by Access for
        browser-based dashboard users. Also accepted via `CF_Authorization`
        cookie.

    S3SigV4Auth:
      type: apiKey
      in: header
      name: Authorization
      description: >
        AWS Signature Version 4 authentication. S3 clients (rclone, boto3,
        aws-cli) send this automatically. Credentials are issued via
        `POST /admin/s3/credentials` (access key prefix `GK`). Also
        supports presigned URL authentication via query parameters.

  # ─── Reusable schemas ──────────────────────────────────────────────────
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [success, errors]
      properties:
        success:
          type: boolean
          const: false
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ApiError'
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
        message:
          type: string

    # --- Bulk operation results ---
...
    PurgeBody:
      type: object
      description: >
        Exactly one of `files`, `hosts`, `tags`, `prefixes`, or
        `purge_everything` must be present.
      properties:
        files:
          type: array
          items:
            oneOf:
              - type: string
                description: Plain URL string.
              - type: object
                required: [url]
                properties:
                  url:
                    type: string
                  headers:
                    type: object
                    additionalProperties:
                      type: string
                    description: >
                      Custom cache key headers (e.g. CF-Device-Type).
          description: Purge by URL(s). Max 500 items (configurable).
        hosts:
          type: array
          items:
            type: string
          description: Purge by hostname(s).
        tags:
          type: array
          items:
            type: string
          description: Purge by cache tag(s).
        prefixes:
          type: array
          items:
            type: string
          description: Purge by URL prefix(es).
        purge_everything:
          type: boolean
          const: true
          description: Purge all cached content in the zone.

    # --- Analytics ---
...
    GatewayConfig:
      type: object
      required:
        - bulk_rate
        - bulk_bucket_size
        - bulk_max_ops
        - single_rate
        - single_bucket_size
        - single_max_ops
        - key_cache_ttl_ms
        - retention_days
      properties:
        bulk_rate:
          type: number
          description: Bulk purge rate limit (req/sec).
        bulk_bucket_size:
          type: number
          description: Bulk purge token bucket capacity.
        bulk_max_ops:
          type: number
          description: Max operations per bulk purge request.
        single_rate:
          type: number
          description: Single-file purge rate limit (URLs/sec).
        single_bucket_size:
          type: number
          description: Single-file purge token bucket capacity.
        single_max_ops:
          type: number
          description: Max URLs per single-file purge request.
        key_cache_ttl_ms:
          type: number
          description: API key cache TTL in milliseconds.
        retention_days:
          type: number
          description: Analytics event retention in days.
    ConfigOverride:
      type: object
      required: [key, value, updated_at]
      properties:
        key:
          type: string
        value:
          type: string
        updated_at:
          type: integer
          description: Unix timestamp (ms).
        updated_by:
          type: ['string', 'null']
    ConfigResponse:
      type: object
      required: [config, overrides, defaults]
      properties:
        config:
          $ref: '#/components/schemas/GatewayConfig'
        overrides:
          type: array
          items:
            $ref: '#/components/schemas/ConfigOverride'
        defaults:
          type: object
          additionalProperties:
            type: number

  # ─── Reusable parameters ─────────────────────────────────────────────────
  parameters:
...
    ZoneIdPath:
      name: zoneId
      in: path
      required: true
      schema:
        type: string
        pattern: '^[a-f0-9]{32}$'
      description: Cloudflare zone ID.
...
paths:
...
  /v1/zones/{zoneId}/purge_cache:
    post:
      operationId: purgeCache
      summary: Purge cache
      description: >
        Proxy to Cloudflare's purge API with IAM policy authorization,
        token-bucket rate limiting, and isolate+DO request collapsing.
        Non-200 responses from the upstream Cloudflare API (including
        429, 500, etc.) are passed through with their original status
        code and body.
      tags: [Purge]
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/ZoneIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PurgeBody'
            examples:
              purge-urls:
                summary: Purge specific URLs
                value:
                  files:
                    - 'https://example.com/styles.css'
                    - 'https://example.com/app.js'
              purge-urls-with-headers:
                summary: Purge URLs with custom cache key headers
                value:
                  files:
                    - url: 'https://example.com/'
                      headers:
                        CF-Device-Type: mobile
              purge-hosts:
                summary: Purge by hostname
                value:
                  hosts: ['cdn.example.com']
              purge-tags:
                summary: Purge by cache tags
                value:
                  tags: ['blog', 'release-v2.1']
              purge-prefixes:
                summary: Purge by URL prefix
                value:
                  prefixes: ['https://example.com/api/']
              purge-everything:
                summary: Purge all cached content
                value:
                  purge_everything: true
      responses:
        '200':
          description: >
            Purge successful. Response body is proxied from Cloudflare's
            upstream API.
          content:
            application/json:
              schema:
                type: object
                description: Cloudflare upstream response (passthrough).
          headers:
            Ratelimit:
              schema:
                type: string
              description: >
                Token bucket state as a Structured Field.
                Format: `"<bucket-name>";r=<remaining>;t=<retry-after-sec>`.
                Example: `"purge-bulk";r=499;t=0`.
              example: '"purge-bulk";r=499;t=0'
            Ratelimit-Policy:
              schema:
                type: string
              description: >
                Rate limit policy as a Structured Field.
                Format: `"<bucket-name>";q=<bucket-size>;w=<window-sec>`.
                Example: `"purge-bulk";q=500;w=10`.
              example: '"purge-bulk";q=500;w=10'
        '400':
          description: Invalid request (bad zone ID, invalid JSON, bad purge body).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: >
            Policy authorization denied. The key's policy does not permit
            this action/resource/condition combination.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ErrorEnvelope'
                  - type: object
                    properties:
                      denied:
                        type: array
                        items:
                          type: string
                        description: >
                          List of items that failed policy evaluation.
                          Present only when specific items were denied.
        '429':
          description: >
            Rate limited — either by the gateway's token bucket or by the
            upstream Cloudflare API. Upstream 429 responses are passed through
            with their original body.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          headers:
            Retry-After:
              schema:
                type: string
              description: Seconds until tokens are available.
            Ratelimit:
              schema:
                type: string
              description: Token bucket state (same format as 200 response).
        '502':
          description: >
            Upstream Cloudflare API request failed (network error, timeout).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

  # ─── Admin: Keys ──────────────────────────────────────────────────────────
...
  /admin/config:
    get:
      operationId: getConfig
      summary: Get gateway configuration
      description: >
        Returns the resolved configuration, active overrides, and defaults.
      tags: [Config]
      security:
        - CloudflareAccess: []
        - AdminKeyAuth: []
      responses:
        '200':
          description: Current configuration.
          content:
            application/json:
              schema:
                type: object
                required: [success, result]
                properties:
                  success:
                    type: boolean
                    const: true
                  result:
                    $ref: '#/components/schemas/ConfigResponse'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

    put:
      operationId: setConfig
      summary: Update gateway configuration
      description: >
        Set one or more config values. Values must be positive finite numbers.
        Only known config keys are accepted.
      tags: [Config]
      security:
        - CloudflareAccess: []
        - AdminKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties:
                type: number
              description: >
                Key-value pairs of config keys to update.
              examples:
                - bulk_rate: 50
                  retention_days: 30
      responses:
        '200':
          description: Configuration updated.
          content:
            application/json:
              schema:
                type: object
                required: [success, result]
                properties:
                  success:
                    type: boolean
                    const: true
                  result:
                    type: object
                    required: [config]
                    properties:
                      config:
                        $ref: '#/components/schemas/GatewayConfig'
        '400':
          description: Validation error (unknown key, invalid value).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...